rename makeUploadForm to just UploadForm; .less file for it
