+++ src/less/app.less
@@ -13,14 +13,11 @@
 @import (less) "components/divisionHeader.less";
 @import (less) "components/institutionStatus.less";
 @import (less) "components/resubmit.less";
+@import (less) "components/uploadForm.less";
 
 body {
   font-family: "AvenirNextLTW01-Regular";
 }
 a {
  cursor: pointer;
-}
-
-input[type="file"] {
-  font-family: "AvenirNextLTW01-Regular";
 }
