+++ src/less/components/uploadForm.less
@@ -1,0 +1,3 @@
+input[type="file"] {
+  font-family: "AvenirNextLTW01-Regular";
+}
